+++ resources/views/dashboard.blade.php
@@ -67,11 +67,15 @@
                 className: 'text-center',
                 name: 'acciones',
                 render: function(data, type, row, meta) {
-                    return `
-                        <a href="{{ url('/') }}/articulos/ver/${row.id}" class="btn btn-primary">Ver</a>
-                        <a href="{{ url('/') }}/articulos/editar/${row.id}" class="btn btn-secondary">Editar</a>
-                        <button onclick="deleteArticle(${row.id})" class="btn btn-danger">Eliminar</button>
-                    `;
+                    if (row.user.id == {{ Auth::id() }}) {
+                        return `
+                            <a href="{{ url('/') }}/articulos/ver/${row.id}" class="btn btn-primary">Ver</a>
+                            <a href="{{ url('/') }}/articulos/editar/${row.id}" class="btn btn-secondary">Editar</a>
+                            <button onclick="deleteArticle(${row.id})" class="btn btn-danger">Eliminar</button>
+                        `;
+                    } else {
+                        return `<a href="{{ url('/') }}/articulos/ver/${row.id}" class="btn btn-primary">Ver</a>`;
+                    }
                 }
             },
         ]
